feat!: add defaultExtent config value
This also controls the extent to which the map zooms when the home button is clicked. Without this change, the home button would zoom to whatever the extent of the map was when the app loaded. Which was an issue when URL params caused the map to load in a non-standard extent.

BREAKING CHANGE: `defaultExtent` needs to be added to your config.